fix linter error
Former-commit-id: cdd5082c03ca851d421175b9ad67c981119b90f4

+++ types/privacytypes.go
@@ -29,6 +29,6 @@ func (ftxos *FTXOsSTXOsInOneTx) SetExpire(tx *Transaction) {
 		ftxos.Expire = Now().Unix() + int64(timeexpire/time.Second+6)
 	} else {
 		// FTXO的过期时间为区块高度时，则用Tx的高度+1
-		ftxos.Expire = int64(expire + 1)
+		ftxos.Expire = expire + 1
 	}
 }
